refactor(llma): move product tool widgets into posthog_ai surface
Move the PostHog product data-tool renderers (query, insight, dashboard, recordings, error-tracking, notebook) from scenes/max/messages/adapters into products/posthog_ai/frontend/components/tool/widgets and have the surface self-register them via ToolCallCard, so every consumer (the /tasks runner, the signals inbox, and Max's sandbox path) renders the widgets instead of a generic JSON card.

Lifts the recordings "accept filters" bar from a maxLogic subscription to a generic RecordingsWidget prop, splits UIPayloadAnswer so the heavy widgets live in the surface while Max keeps the LangGraph-only exports, and exposes VisualizationWidget / RecordingsWidget / ErrorTrackingFiltersWidget / ErrorTrackingIssueCard through api/primitives for the LangGraph consumers.

Generated-By: PostHog Code
Task-Id: 8f11b113-3a6b-4bea-b1ed-a212db1aa00d
