+++ frontend/src/scenes/max/messages/AnnotatedScreenshotView.tsx
