+++ frontend/src/scenes/max/messages/VisualizationArtifact.tsx
@@ -10,8 +10,9 @@ import { Scene } from 'scenes/sceneTypes'
 
 import { ArtifactMessage, VisualizationArtifactContent } from '~/queries/schema/schema-assistant-messages'
 
+import { VisualizationWidget, getArtifactOpenTarget } from 'products/posthog_ai/frontend/api/primitives'
+
 import { MessageStatus } from '../maxLogic'
-import { VisualizationWidget, getArtifactOpenTarget } from './VisualizationWidget'
 
 interface VisualizationArtifactProps {
     message: ArtifactMessage & { status?: MessageStatus }
