+++ products/logs/frontend/components/LogsViewer/LogDetailsModal/Tabs/RelatedErrors/RelatedErrorsTab.tsx
@@ -3,9 +3,10 @@ import { BindLogic, useActions, useValues } from 'kea'
 import { LemonBanner, LemonSkeleton } from '@posthog/lemon-ui'
 
 import { EmptyMessage } from 'lib/components/EmptyMessage/EmptyMessage'
-import { ErrorTrackingIssueCard } from 'scenes/max/messages/ErrorTrackingIssueCard'
 
 import { MaxErrorTrackingIssuePreview } from '~/queries/schema/schema-assistant-error-tracking'
+
+import { ErrorTrackingIssueCard } from 'products/posthog_ai/frontend/api/primitives'
 
 import { RelatedErrorsLogicProps, relatedErrorsLogic } from './relatedErrorsLogic'
