+++ products/posthog_ai/frontend/api/primitives.ts
@@ -75,3 +75,11 @@ export { ContextUsageBar } from '../components/ContextUsageBar'
 export { QuestionField, MultiFieldQuestion, isFieldValid } from '../components/QuestionField'
 export { OptionSelector } from '../components/OptionSelector'
 export type { Option } from '../components/OptionSelector'
+
+// Product data-tool widgets — the atomic PostHog-entity renderers Max's LangGraph path composes directly
+// (the sandbox path reaches them through the tool registry, not these exports).
+export { VisualizationWidget, getArtifactOpenTarget } from '../components/tool/widgets/VisualizationWidget'
+export type { VisualizationWidgetProps } from '../components/tool/widgets/VisualizationWidget'
+export { RecordingsWidget } from '../components/tool/widgets/RecordingsWidget'
+export { ErrorTrackingFiltersWidget } from '../components/tool/widgets/ErrorTrackingFiltersWidget'
+export { ErrorTrackingIssueCard } from '../components/tool/widgets/ErrorTrackingIssueCard'
