+++ products/posthog_ai/frontend/components/tool/ToolCallCard.tsx
@@ -1,3 +1,8 @@
+// Side-effect: registers the surface's product data-tool renderers (insight/dashboard/recordings/etc.)
+// into the shared toolRegistry. Placed at the render chokepoint so every surface that renders a tool
+// card (the /tasks runner, the signals inbox, Max's sandbox path) resolves those widgets.
+import './widgets/registerDataToolRenderers'
+
 import { Suspense, memo } from 'react'
 
 import type { ToolCallMessage } from 'products/posthog_ai/frontend/types/toolTypes'
