+++ products/posthog_ai/frontend/components/tool/widgets/CreateNotebookWidget.test.tsx
@@ -1,10 +1,10 @@
-// Side-effect: Max registers its product tool renderers (incl. the notebook keys) into the shared
-// registry. The bare registry no longer knows product keys, so the resolution assertions below need it.
-import './registerMaxToolRenderers'
+// Side-effect: registers the surface's product data-tool renderers (incl. the notebook keys) into the
+// shared registry. The bare registry no longer knows product keys, so the resolution assertions below need it.
+import './registerDataToolRenderers'
 
 import { lookupToolRenderer, toolRegistry } from 'products/posthog_ai/frontend/api/tools'
+import type { ToolCallMessage } from 'products/posthog_ai/frontend/types/toolTypes'
 
-import type { ToolCallMessage } from '../../maxTypes'
 import { extractNotebook } from './CreateNotebookWidget'
 
 function toolMessage(rawOutput: unknown, innerInput?: Record<string, unknown>): ToolCallMessage {
