+++ products/posthog_ai/frontend/components/tool/widgets/ErrorTrackingIssueCard.tsx
