+++ products/posthog_ai/frontend/components/tool/widgets/ErrorTrackingWidget.tsx
@@ -1,6 +1,7 @@
-import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend/api/tools'
-
-import { ErrorTrackingFiltersWidget } from '../UIPayloadAnswer'
+import { DataToolRow } from '../DataToolRow'
+import { GenericMcpToolRenderer } from '../GenericMcpToolRenderer'
+import type { ToolRendererProps } from '../toolRegistry'
+import { ErrorTrackingFiltersWidget } from './ErrorTrackingFiltersWidget'
 import { extractErrorTrackingResponse } from './extractors'
 
 /**
