+++ products/posthog_ai/frontend/components/tool/widgets/QueryWidget.tsx
@@ -1,7 +1,8 @@
-import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend/api/tools'
-
-import { VisualizationWidget } from '../VisualizationWidget'
+import { DataToolRow } from '../DataToolRow'
+import { GenericMcpToolRenderer } from '../GenericMcpToolRenderer'
+import type { ToolRendererProps } from '../toolRegistry'
 import { extractQueryResult } from './extractors'
+import { VisualizationWidget } from './VisualizationWidget'
 
 /**
  * Renders the MCP query-wrapper tools (query-trends, query-funnel, ...) through
