+++ products/posthog_ai/frontend/components/tool/widgets/SearchSessionRecordingsWidget.tsx
@@ -1,7 +1,8 @@
-import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend/api/tools'
-
-import { RecordingsWidget } from '../UIPayloadAnswer'
+import { DataToolRow } from '../DataToolRow'
+import { GenericMcpToolRenderer } from '../GenericMcpToolRenderer'
+import type { ToolRendererProps } from '../toolRegistry'
 import { extractRecordingFilters } from './extractors'
+import { RecordingsWidget } from './RecordingsWidget'
 
 /**
  * Session-recording search / filter tool calls. The resolved `RecordingUniversalFilters` come back
